Update: Error messages in Auth pages
