+++ frontend/src/pages/SignUp/SignUp.jsx
@@ -10,54 +10,17 @@ import {
 } from '@/styledComponents'
 import { useAuth, useMyForm } from '@/hooks'
 import { Header } from '@/components/Common/Header'
-
-const customErrors = (values) => {
-  const newErrors = {}
-
-  // Name errors
-  if (values.name === '') newErrors.name = 'Please complete this field'
-  else if (values.name.length < 2) newErrors.name = 'Min length 2 characters'
-  else newErrors.name = ''
-
-  // Surname errors
-  if (values.surname === '') newErrors.surname = 'Please complete this field'
-  else if (values.surname.length < 2) newErrors.surname = 'Min length 2 characters'
-  else newErrors.surname = ''
-
-  // Username errors
-  if (values.username === '') newErrors.username = 'Please complete this field'
-  else if (values.username.length < 8) newErrors.username = 'Min length 8 characters'
-  else newErrors.username = ''
-
-  // Email errors
-  if (values.email === '') newErrors.email = 'Please complete this field'
-  else if (values.email.length < 8) newErrors.email = 'Min length 8 characters'
-  else newErrors.email = ''
-
-  // Password errors
-  if (values.password === '') newErrors.password = 'Please complete this field'
-  else if (values.password.length < 8) newErrors.password = 'Min length 8 characters'
-  else newErrors.password = ''
-
-  return newErrors
-}
+import { customErrors } from './config'
 
 export default function SignUp () {
-  const [error, setError] = useState('')
+  const [error, setError] = useState(null)
   const { signup, logout, isLogged, loading } = useAuth()
-  const { getAttributes, submit, validate } = useMyForm({
-    name: '',
-    surname: '',
-    username: '',
-    email: '',
-    password: ''
-  })
+  const { getAttributes, submit, validate } = useMyForm()
 
   useEffect(() => {
     isLogged && logout({ redirect: false })
   }, [])
 
-  // useValidate(customErrors)
   validate(customErrors)
 
   const handleSubmit = async (values) => {
@@ -134,9 +97,11 @@ export default function SignUp () {
             </SubmitButton>
           </FormGridItem>
           {Boolean(error) && <FormGridItem>
-            <Typography color='error'>
-              Username or password incorrects
+            {error.map((item, index) => (
+            <Typography key={index} align='left' color='error'>
+              - {item}
             </Typography>
+            ))}
           </FormGridItem>}
         </FormGridContainer>
       </AuthenticateFormContainer>
