+++ frontend/src/pages/SignUp/config/index.js
@@ -1,0 +1,1 @@
+export { default as customErrors } from './errors'
